requirements.txt
flask==2.0.1
werkzeug==2.0.2
requests==2.28.1
python-dotenv==0.21.0
markdown==3.4.1
pygments==2.13.0
Pillow==9.3.0  # For image processing